auth-service/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
python-consul==1.1.0
SQLAlchemy==2.0.23